update readme and restructure
