+++ README.md
@@ -7,12 +7,19 @@ This repository is under development. Expect anything in it to change rapidly.
 
 You have now successfully installed the repository!
 
+### Running the random agents:
+1. Run `source .venv/bin/activate` to enter the virtual environement
+2. Navigate to the environment you want to test, either `swapenv` or `colorenv`
+3. Run the corresponding python script, for example: `python random_swap_agent.py`
+
 ### Running the examples:
 1. Run `source .venv/bin/activate` to enter the virtual environement
-2. Run `python random_agent.py` to get the average reward of a random agent in the coloring environment
-3. Run `python test_tune.py` to see Rllib train a PPO agent on CartPole and perform a grid search for the best learning rate
-4. Run `python color_tune.py` to train a DQN agent on a 3x3 coloring environment with 2 colors
-5. Run `deactivate` to exit the virtual environment
+2. Run `python test_tune.py` to see Rllib train a PPO agent on CartPole and perform a grid search for the best learning rate
+3. Run `python color_tune.py` to train a DQN agent on a 3x3 coloring environment with 2 colors
+4. Run `python swap_tune.py` to train a PPO agent on a 3x3 swap environment with 2 colors
 
 ### Visualization:
 After running `python test_tune.py` or `python color_tune.py` you can run `tensorboard --logdir=~/ray_results/` in another window to visualize training in your browser
+
+### Playing the environment yourself:
+Currently this is only supported by the `swapenv`. You can play by navigating to the `swapenv` directory an running `python play.py`
