+++ colorenv/random_coloring_agent.py
@@ -1,11 +1,11 @@
 import gym
 import statistics
 import os
-from colorenv.colorenv import ColorEnv
+from colorenv import ColorEnv
 
 rewards = []
 num_iterations = 100000
-path_to_config = os.getcwd() + "/colorenv/config.yaml"
+path_to_config = os.getcwd() + "/config.yaml"
 env = ColorEnv(path_to_config)
 observation = env.reset()
