+++ swapenv/random_swap_agent.py
@@ -1,11 +1,11 @@
 import gym
 import statistics
 import os
-from swapenv.swapenv import SwapEnv
+from swapenv import SwapEnv
 
 rewards = []
 num_iterations = 100000
-path_to_config = os.getcwd() + "/swapenv/config.yaml"
+path_to_config = os.getcwd() + "/config.yaml"
 env = SwapEnv(path_to_config)
 observation = env.reset()
